# [Storage] Optimize speed & memory for generating db keys (BadgerDB and Pebble)