--- a/storage/badger/operation/prefix.go
+++ b/storage/badger/operation/prefix.go
@@ -1,10 +1,7 @@
 package operation
 
 import (
-	"encoding/binary"
-	"fmt"
-
-	"github.com/onflow/flow-go/model/flow"
+	op "github.com/onflow/flow-go/storage/operation"
 )
 
 const (
@@ -115,36 +112,10 @@ const (
 	codeEpochEmergencyFallbackTriggered = 255
 )
 
-func makePrefix(code byte, keys ...interface{}) []byte {
-	prefix := make([]byte, 1)
-	prefix[0] = code
-	for _, key := range keys {
-		prefix = append(prefix, b(key)...)
-	}
-	return prefix
+func makePrefix(code byte, keys ...any) []byte {
+	return op.MakePrefix(code, keys...)
 }
 
-func b(v interface{}) []byte {
-	switch i := v.(type) {
-	case uint8:
-		return []byte{i}
-	case uint32:
-		b := make([]byte, 4)
-		binary.BigEndian.PutUint32(b, i)
-		return b
-	case uint64:
-		b := make([]byte, 8)
-		binary.BigEndian.PutUint64(b, i)
-		return b
-	case string:
-		return []byte(i)
-	case flow.Role:
-		return []byte{byte(i)}
-	case flow.Identifier:
-		return i[:]
-	case flow.ChainID:
-		return []byte(i)
-	default:
-		panic(fmt.Sprintf("unsupported type to convert (%T)", v))
-	}
+func keyPartToBinary(v any) []byte {
+	return op.AppendPrefixKeyPart(nil, v)
 }