This optimization was found while getting familiar with storage code base related to BadgerDB and Pebble.

`MakePrefix()` is used to generate BadgerDB and Pebble keys. It concatenates prefix with multiple key parts in binary format, which can create many allocations during concatenation.

Benchmark `MakePrefixWithID` concatenates prefix with Identifier, which is common use case.

```
                    │ original.log  │            optimized.log             │
                    │    sec/op     │    sec/op     vs base                │
MakePrefixWithID-12   145.55n ± 11%   96.11n ± 10%  -33.97% (p=0.000 n=20)

                    │ original.log │           optimized.log            │
                    │     B/op     │    B/op     vs base                │
MakePrefixWithID-12    113.00 ± 0%   80.00 ± 0%  -29.20% (p=0.000 n=20)

                    │ original.log │           optimized.log            │
                    │  allocs/op   │ allocs/op   vs base                │
MakePrefixWithID-12     4.000 ± 0%   2.000 ± 0%  -50.00% (p=0.000 n=20)
```

Benchmark `MakePrefixWithChainIDAndNum` concatenates prefix with `ChainID` and number (uint64), which is used to index ClusterBlockHeight.
```
                               │ original.log │            optimized.log             │
                               │    sec/op    │    sec/op     vs base                │
MakePrefixWithChainIDAndNum-12   145.90n ± 5%   58.93n ± 11%  -59.61% (p=0.000 n=20)

                               │ original.log │           optimized.log            │
                               │     B/op     │    B/op     vs base                │
MakePrefixWithChainIDAndNum-12     96.00 ± 0%   40.00 ± 0%  -58.33% (p=0.000 n=20)

                               │ original.log │           optimized.log            │
                               │  allocs/op   │ allocs/op   vs base                │
MakePrefixWithChainIDAndNum-12     6.000 ± 0%   2.000 ± 0%  -66.67% (p=0.000 n=20)
```

This commit also refactors duplicate copies of `MakePrefix()`:
- remove unused storage/pebble/operation/prefix.go
- update badger/operation/prefix.go to call optimized MakePrefix() in storage/operation/prefix.go